Added the constants in Json, that can be used from Typescript, inside a Switch.
I spent the last few days to see if we could make references to strings inside json schema.
The answer is yes but.
Yes, you can create properties:
"referenceThis": { "const": "prop_is_picked_up" }
And then reference it with
“$ref” : “#/referenceThis”
And it will work…
But if you use these as:
In if/then matching - it won’t match the target
In an enum - then both
Enum rule enforcement - I forget what this had, but it doesn’t work.
Enum autocomplete puts {“$ref” : “blah”} in the autocomplete dropdown, instead of the term ← this is probably the fault of the UI
So instead, I’ve done i